dg-backend/requirements.txt
mysql-connector-python
langchain-ollama
chromadb
python-dotenv
